feat: update Profile form
